+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuDivider.kt
@@ -1,0 +1,21 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package mozilla.components.browser.menu.item
+
+import android.view.View
+import mozilla.components.browser.menu.BrowserMenu
+import mozilla.components.browser.menu.BrowserMenuItem
+import mozilla.components.browser.menu.R
+
+/**
+ * A browser menu item to display a horizontal divider.
+ */
+class BrowserMenuDivider : BrowserMenuItem {
+    override var visible: () -> Boolean = { true }
+
+    override fun getLayoutResource() = R.layout.mozac_browser_menu_item_divider
+
+    override fun bind(menu: BrowserMenu, view: View) = Unit
+}